chameleon-api/requirements.txt
Flask==3.1.1
flask-limiter==3.12
flask-apscheduler==1.13.1
Flask-Cors==5.0.0
gmssl==3.2.2
Pillow==11.3.0
requests==2.32.4
PyJWT==2.10.1
Werkzeug==3.1.3
redis>=6.3.0
dashscope>=1.24.1